+++ src/main/java/TimeIsGold/TimeIsGold/api/login/dto/LoginRequestDto.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Data
 public class LoginRequestDto {
+
     public String id;
     public String pw;
 }
