+++ src/main/java/TimeIsGold/TimeIsGold/configuration/AppConfig.java
@@ -1,4 +1,4 @@
-package TimeIsGold.TimeIsGold;
+package TimeIsGold.TimeIsGold.configuration;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
