+++ src/main/java/TimeIsGold/TimeIsGold/configuration/WebConfig.java
@@ -1,0 +1,17 @@
+package TimeIsGold.TimeIsGold.configuration;
+
+import TimeIsGold.TimeIsGold.interceptor.LoginInterceptor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new LoginInterceptor())
+                .addPathPatterns("/**")
+                .excludePathPatterns("/", "/login", "/logout");
+    }
+}
